% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{align_gene_expression_metadata}
\alias{align_gene_expression_metadata}
\title{Match gene expression data with metadata}
\usage{
align_gene_expression_metadata(dataList = NULL, metadata = NULL, counts = NULL)
}
\arguments{
\item{dataList}{A list containing metadata and gene expression data (counts) or individual metadata and counts as separate arguments}

\item{metadata}{Optional argument. Metadata, required if counts is provided separately from dataList.}

\item{counts}{Optional argument. Gene expression data (counts), required if metadata is provided separately from dataList.}
}
\value{
A list containing the matched counts and metadata
}
\description{
Match gene expression data with metadata
}
