% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CaExpressionCounts}
\alias{CaExpressionCounts}
\title{Gene Expression Counts
Gene expression counts for 28891 gene and 197 samples}
\format{
matrix with 28891 rows and 197 columns
}
\source{
https://www.ncbi.nlm.nih.gov
}
\usage{
CaExpressionCounts
}
\description{
Gene Expression Counts
Gene expression counts for 28891 gene and 197 samples
}
\examples{
data(CaExpressionCounts)
}
\keyword{datasets}
