% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CaExpressionMetadata}
\alias{CaExpressionMetadata}
\title{Gene Expression Metadata
Metadata for 197 samples across 14 different studies and 6 different stress conditions}
\format{
data.frame with 197 rows and 6 columns
}
\source{
https://www.ncbi.nlm.nih.gov
}
\usage{
CaExpressionMetadata
}
\description{
Gene Expression Metadata
Metadata for 197 samples across 14 different studies and 6 different stress conditions
}
\examples{
data(CaExpressionMetadata)
}
\keyword{datasets}
