% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CaExpressionSignificance}
\alias{CaExpressionSignificance}
\title{Significance of gene expression
Significance of gene expression for 28891 genes}
\format{
data.frame "geneid", "pvalue", "FDR", "logfc", "exp.direction", "study", "experiment"
}
\usage{
CaExpressionSignificance
}
\description{
Significance of gene expression
Significance of gene expression for 28891 genes
}
\examples{
data(CaExpressionSignificance)
}
\keyword{datasets}
