% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CaGenomeAnnotation}
\alias{CaGenomeAnnotation}
\title{Gene annotation data
Annotation data for 24681 genes}
\format{
data.frame with 24681 rows and 6 columns
}
\usage{
CaGenomeAnnotation
}
\description{
Gene annotation data
Annotation data for 24681 genes
}
\examples{
data(CaGenomeAnnotation)
}
\keyword{datasets}
