% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CaGWAS}
\alias{CaGWAS}
\title{GWAS data
GWAS data for 1052 genes}
\format{
An object of class \code{data.frame} with 7816 rows and 15 columns.
}
\usage{
CaGWAS
}
\description{
GWAS data
GWAS data for 1052 genes
}
\keyword{datasets}
