% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CaProteinEnrichment}
\alias{CaProteinEnrichment}
\title{Protein enrichment data
Protein enrichment data for 24254 genes}
\format{
data.frame with 24254 rows and 5 columns
}
\usage{
CaProteinEnrichment
}
\description{
Protein enrichment data
Protein enrichment data for 24254 genes
}
\examples{
data(CaProteinEnrichment)
}
\keyword{datasets}
