% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CaProteinInteractionNetwork}
\alias{CaProteinInteractionNetwork}
\title{Protein interaction network
Protein interaction network for 24681 genes}
\format{
data.frame 9 columns
}
\usage{
CaProteinInteractionNetwork
}
\description{
Protein interaction network
Protein interaction network for 24681 genes
}
\examples{
data(CaProteinInteractionNetwork)
}
\keyword{datasets}
