% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CaProteinSequenceInfo}
\alias{CaProteinSequenceInfo}
\title{Protein sequence information
Protein sequence information for 24681 genes}
\format{
data.frame with 24681 rows and 5 columns
}
\usage{
CaProteinSequenceInfo
}
\description{
Protein sequence information
Protein sequence information for 24681 genes
}
\examples{
data(CaProteinSequenceInfo)
head(CaProteinSequenceInfo)
}
\keyword{datasets}
