% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_gene_expression.R
\name{filter_gene_expression}
\alias{filter_gene_expression}
\title{Filter Gene Expression Data
Filter gene expression data based minimum expression level}
\usage{
filter_gene_expression(data)
}
\arguments{
\item{data}{gene expression matrix}

\item{minexp}{minimum expression level}
}
\value{
filtered gene expression matrix
}
\description{
Filter Gene Expression Data
Filter gene expression data based minimum expression level
}
\examples{
data("CaExpressionCounts")
}
