% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{gene_pheno_GWASmatrix_byGeneID}
\alias{gene_pheno_GWASmatrix_byGeneID}
\title{Get gene-phenotype matrix from GWAS data by gene ids}
\usage{
gene_pheno_GWASmatrix_byGeneID(geneids, pvalue = 0.05, phenotype = NULL)
}
\arguments{
\item{geneids}{The gene ids to search for}

\item{pvalue}{The pvalue threshold}

\item{phenotype}{The phenotype to filter the data}
}
\value{
gene Pheno matrix from GWAS
}
\description{
Get gene-phenotype matrix from GWAS data by gene ids
}
