% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{gene_pheno_GWASmatrix_byGOKeyword}
\alias{gene_pheno_GWASmatrix_byGOKeyword}
\title{Get gene expression data using GO term keywords}
\usage{
gene_pheno_GWASmatrix_byGOKeyword(keyword, pvalue = 0.05, phenotype = NULL)
}
\arguments{
\item{phenotype}{The phenotype to filter the data}

\item{keywords}{The keywords to search for}

\item{pvlaue}{The pvalue threshold}
}
\value{
gene Pheno matrix from GWAS
}
\description{
Get gene expression data using GO term keywords
}
