% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{generate_heatmap_by_go_term_keyword}
\alias{generate_heatmap_by_go_term_keyword}
\title{Generate a heatmap for a given GO term keyword}
\usage{
generate_heatmap_by_go_term_keyword(
  keyword,
  study = NULL,
  experiment = NULL,
  outfolder = "",
  plotwidth = 10,
  plotheight = 10
)
}
\arguments{
\item{keyword}{The keyword to search for}

\item{study}{The study to filter the gene expression data}

\item{experiment}{The experiment to filter the gene expression data}

\item{outfolder}{The folder to save the heatmap and data files}

\item{plotwidth}{The width of the plot (default: 10)}

\item{plotheight}{The height of the plot (default: 10)}
}
\description{
Generate a heatmap for a given GO term keyword
}
\examples{
generate_heatmap_by_go_term_keyword(c("Flowering"),study="Salinity", experiment=c("PRJNA232700","PRJNA288473"), outfolder="heatmaps", plotwidth=10, plotheight=20)
}
