% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_data_report}
\alias{get_data_report}
\title{Get report for data}
\usage{
get_data_report(data)
}
\arguments{
\item{data}{The data to get the report for}
}
\value{
A report for the data
}
\description{
Get report for data
}
\examples{
get_report(data)
get_report(get_study_data("Salinity"))
}
