% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_gene_annotation_bygeneid}
\alias{get_gene_annotation_bygeneid}
\title{Get gene annotation data for a specific geneid}
\usage{
get_gene_annotation_bygeneid(geneids = c())
}
\arguments{
\item{geneids}{The gene or genes to get the annotation for}
}
\value{
Gene annotation data
}
\description{
Get gene annotation data for a specific geneid
}
\examples{
get_gene_annotation(c("GRX3", "HIS2A","HSFA1B","HSFA2","HSFA3"))
get_gene_annotation()
}
