% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{get_gene_enrichment_bygeneid}
\alias{get_gene_enrichment_bygeneid}
\title{Get gene enrichment data for a specific geneid}
\usage{
get_gene_enrichment_bygeneid(geneids)
}
\arguments{
\item{geneids}{The gene or genes to get the enrichment for}
}
\value{
Gene enrichment data
}
\description{
Get gene enrichment data for a specific geneid
}
