% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_gene_express_bygeneid}
\alias{get_gene_express_bygeneid}
\title{Get gene expression counts for a specific geneid}
\usage{
get_gene_express_bygeneid(geneids, gene_express_count = NULL)
}
\arguments{
\item{genenames}{The gene or genes to get the counts for}
}
\value{
Expression Count matrix for the gene or genes
}
\description{
Get gene expression counts for a specific geneid
}
\examples{
get_gene_express_bygeneid(c("GRX3", "HIS2A","HSFA1B","HSFA2","HSFA3"))
}
