% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{get_gene_express_bygoid}
\alias{get_gene_express_bygoid}
\title{Get gene expression data using GO term id}
\usage{
get_gene_express_bygoid(go_id, gene_express_count = NULL)
}
\arguments{
\item{go_id}{The GO term id to get the data for}

\item{gene_express_count}{The gene expression count}
}
\value{
Gene expression data
}
\description{
Get gene expression data using GO term id
}
\examples{
get_gene_express_bygoid("GO:0071704","GO:1901564","GO:0005622","GO:0005634","GO:0043226","GO:0043227","GO:0043229","GO:0043231","GO:0110165")
}
