% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_gene_express_bysamples}
\alias{get_gene_express_bysamples}
\title{Get gene expression counts for a specific samples}
\usage{
get_gene_express_bysamples(samples)
}
\arguments{
\item{samples}{The samples to get the counts for}
}
\value{
Expression Count matrix for the samples
}
\description{
Get gene expression counts for a specific samples
}
\examples{
get_gene_express_bysamples(c("SRR10331638", "SRR10331639", "SRR10331640", "SRR10331641", "SRR10331642", "SRR10331643", "SRR10331644", "SRR10331645"))
}
