% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{get_gene_express_gokeywords}
\alias{get_gene_express_gokeywords}
\title{Get gene expression data using GO term keywords}
\usage{
get_gene_express_gokeywords(keywords, gene_expression_count = NULL)
}
\arguments{
\item{keywords}{The keywords to search for}

\item{gene_expression_count}{The gene expression count}
}
\value{
Gene expression data
}
\description{
Get gene expression data using GO term keywords
}
\examples{
get_gene_express_gotermkeywords(c("targeting"))
get_gene_express_gotermkeywords(c("targeting"), CaExpressionCounts)
}
