% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{get_gene_logfc}
\alias{get_gene_logfc}
\title{Get gene expression significance data}
\usage{
get_gene_logfc(
  genesSample,
  pvalue = 1,
  fdr = 1,
  study = "all",
  experiment = "all"
)
}
\arguments{
\item{genesSample}{The gene or genes to get the logfc for}

\item{pvalue}{The pvalue threshold}

\item{fdr}{The fdr threshold}

\item{study}{The study to get the logfc for}

\item{experiment}{The experiment to get the logfc for}
}
\value{
Gene logfc data
}
\description{
Get gene expression significance data
}
