% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{get_gene_logfc_bycategory}
\alias{get_gene_logfc_bycategory}
\title{Get the mean logfc for gene list across categories}
\usage{
get_gene_logfc_bycategory(geneids, pvalue = 1, fdr = 1, category = "study")
}
\arguments{
\item{geneids}{The gene or genes to get the logfc for}

\item{pvalue}{The pvalue threshold}

\item{fdr}{The fdr threshold}

\item{category}{The category to get the logfc for}
}
\description{
Get the mean logfc for gene list across categories
}
