% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{get_geneids_bygotermkeywords}
\alias{get_geneids_bygotermkeywords}
\title{Get gene ids by GO term keywords}
\usage{
get_geneids_bygotermkeywords(keywords)
}
\arguments{
\item{keywords}{The keywords to search for}
}
\value{
Gene ids
}
\description{
Get gene ids by GO term keywords
}
\examples{
get_geneids_bygotermkeywords(c("targeting"))
}
