% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_goids_bykewords}
\alias{get_goids_bykewords}
\title{Get GO ids by keywords}
\usage{
get_goids_bykewords(keywords)
}
\arguments{
\item{keywords}{The keywords to search for}
}
\value{
GO ids
}
\description{
Get GO ids by keywords
}
\examples{
get_goids_bykewords(c("targeting"))
}
