% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{get_gwas_data}
\alias{get_gwas_data}
\title{Get GWAS data for specific genes}
\usage{
get_gwas_data(geneids, pvalue = 0.05, phenotype = NULL)
}
\arguments{
\item{geneids}{The genes to get the GWAS data for}

\item{pvalue}{The p-value threshold to use}

\item{phenotype}{The phenotype to get the GWAS data for}
}
\value{
GWAS data for the genes
}
\description{
Get GWAS data for specific genes
}
