% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata for a specific samples, group, experiment, or study}
\usage{
get_metadata(samples = NULL, group = NULL, experiment = NULL, study = NULL)
}
\arguments{
\item{samples}{The samples to get the metadata for}

\item{group}{The group to get the metadata for}

\item{experiment}{The experiment to get the metadata for}

\item{study}{The study to get the metadata for}
}
\value{
Metadata for the samples, group, experiment, or study
}
\description{
Get Metadata for a specific samples, group, experiment, or study
}
\examples{
get_metadata(samples = c("SRR10331638", "SRR10331639", "SRR10331640", "SRR10331641", "SRR10331642", "SRR10331643", "SRR10331644", "SRR10331645"))
}
