% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_protein_interaction_table}
\alias{get_protein_interaction_table}
\title{Get the protein interaction network table for specific genes}
\usage{
get_protein_interaction_table(genes, only_interacting = FALSE)
}
\arguments{
\item{genes}{a character vector of gene names}
}
\value{
a data frame with protein interaction network information
}
\description{
Get the protein interaction network table for specific genes
}
\examples{
get_protein_interaction_table(c("LOC101501102","LOC101507179"))
}
