% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_samples_data}
\alias{get_samples_data}
\title{Get the data for a specific set of samples}
\usage{
get_samples_data(samples = c())
}
\arguments{
\item{samples}{The samples to get the data for}
}
\value{
A list with the counts and metadata for the samples
}
\description{
Get the data for a specific set of samples
}
\examples{
get_samples_data(c("SRR10331638", "SRR10331639", "SRR10331640", "SRR10331641", "SRR10331642", "SRR10331643", "SRR10331644", "SRR10331645")
}
