% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{get_study_data}
\alias{get_study_data}
\title{Get the data for a specific stress study}
\usage{
get_study_data(study = "all")
}
\arguments{
\item{stress}{The stress study to get the data for}
}
\value{
A list with the counts and metadata for the study
}
\description{
Get the data for a specific stress study
}
\examples{
get_stress_data("Salinity")
}
