% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{loadCaExpressionMetadata}
\alias{loadCaExpressionMetadata}
\title{Load CaExpressionMetadata}
\usage{
loadCaExpressionMetadata()
}
\description{
Load CaExpressionMetadata
}
\examples{
loadCaExpressionMetadata()
}
