% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{loadCaProteinEnrichment}
\alias{loadCaProteinEnrichment}
\title{Load CaProteinEnrichment}
\usage{
loadCaProteinEnrichment()
}
\description{
Load CaProteinEnrichment
}
\examples{
loadCaProteinEnrichment()
}
