% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{loadCaProteinInteractionNetwork}
\alias{loadCaProteinInteractionNetwork}
\title{Load the gene expression count data}
\usage{
loadCaProteinInteractionNetwork()
}
\description{
Load the gene expression count data
}
