% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{plot_gene_expression_heatmap}
\alias{plot_gene_expression_heatmap}
\title{Plot gene expression heatmap}
\usage{
plot_gene_expression_heatmap(
  geneExp,
  plot_height = 10,
  plot_width = 10,
  file = "gene_expression_heatmap.pdf"
)
}
\arguments{
\item{geneExp}{Gene expression data}

\item{plot_height}{The height of the plot}

\item{plot_width}{The width of the plot}

\item{file}{The file to save the plot}
}
\value{
The plot
}
\description{
Plot gene expression heatmap
}
