% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{plot_gene_pheno_GWASmatrix_byGOKeyword}
\alias{plot_gene_pheno_GWASmatrix_byGOKeyword}
\title{Plot gene-phenotype heatmap from GWAS data}
\usage{
plot_gene_pheno_GWASmatrix_byGOKeyword(
  gene_pheno_matrix = NULL,
  outfolder = NULL,
  plotwidth = 10,
  plotheight = 10,
  keyword = NULL,
  pvalue = 0.05,
  phenotype = NULL
)
}
\arguments{
\item{gene_pheno_matrix}{The gene-phenotype matrix}

\item{outfolder}{The folder to save the plot}

\item{plotwidth}{The width of the plot (default: 10)}

\item{plotheight}{The height of the plot (default: 10)}

\item{keyword}{The keyword to search for}

\item{pvalue}{The pvalue threshold}

\item{phenotype}{The phenotype to filter the data}
}
\description{
Plot gene-phenotype heatmap from GWAS data
}
