% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{plot_protein_interaction_network}
\alias{plot_protein_interaction_network}
\title{Plot a protein interaction network without gene expression colors}
\usage{
plot_protein_interaction_network(
  network_table,
  targetCol = "score",
  title = "Protein Interaction Network",
  file = NULL,
  plot_width = 10,
  plot_height = 10
)
}
\arguments{
\item{network_table}{a data frame with protein interaction network information}

\item{targetCol}{the column name for the target score}

\item{title}{the title for the plot}
}
\description{
Plot a protein interaction network without gene expression colors
}
