% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{plot_protein_interaction_network_with_gene_expression}
\alias{plot_protein_interaction_network_with_gene_expression}
\title{Plot a protein interaction network with gene expression data}
\usage{
plot_protein_interaction_network_with_gene_expression(
  network_table,
  geneExp,
  targetCol = "score",
  title = "Protein Interaction Network",
  file = NULL,
  plot_width = 10,
  plot_height = 10,
  minnodes = 3
)
}
\arguments{
\item{network_table}{The protein interaction network table}

\item{geneExp}{The gene expression data}

\item{targetCol}{The column in geneExp to use for coloring nodes}

\item{title}{The title of the plot}

\item{file}{The file path to save the plot (optional)}

\item{plot_width}{The width of the plot (default: 10)}

\item{plot_height}{The height of the plot (default: 10)}

\item{minnodes}{The minimum number of connections a node must have to be included in the plot (default: 3)}
}
\value{
A plot of the protein interaction network with gene expression data
}
\description{
Plot a protein interaction network with gene expression data
}
