% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{.onAttach}
\alias{.onAttach}
\title{Loading required data}
\usage{
.onAttach(libname, pkgname)
}
\description{
ChickpeaOmicsR is a package that provides gene expression data for chickpea under different stress conditions.
}
