% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{filter_gene_expression}
\alias{filter_gene_expression}
\title{Filter gene expression data based minimum expression level}
\usage{
filter_gene_expression(data, min_expression = 10)
}
\arguments{
\item{data}{gene expression matrix}
}
\value{
filtered gene expression matrix
}
\description{
Filter gene expression data based minimum expression level
}
\examples{
data("CaExpressionCounts")
}
