% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_geneids_bygoids}
\alias{get_geneids_bygoids}
\title{Get gene ids by GO ids}
\usage{
get_geneids_bygoids(goids)
}
\arguments{
\item{goids}{The GO ids to search for}
}
\value{
Gene ids
}
\description{
Get gene ids by GO ids
}
\examples{
get_geneids_bygoids(c("GO:0071704","GO:1901564","GO:0005622","GO:0005634","GO:0043226","GO:0043227","GO:0043229","GO:0043231","GO:0110165"))
}
